verify packaged desktop runtime in release CI
